refactor : 불필요하게 출력되는 부분 제거
